numberOnly

+++ views/controller/mainController.js
@@ -3,22 +3,26 @@ angular.module('mainAppModule', ['loginModule'])
         var refresh = function() {
             $http.get('/itemList').success(function(response){
                 $scope.itemList = response;
-                $scope.item = {
-                    name:"", price:"0", qty:"0"
-                };
+                refreshInput();
             });
         };
         refresh();
-
+        var refreshInput = function() {
+            $scope.item = {
+                name:"", price:"0", qty:"0"
+            };
+        };
         $scope.addItem = function(){
             $http.post('/itemList', $scope.item).success(function(response){
                 refresh();
             });
         };
-        $scope.deleteItem = function(id){
-            $http.delete('/itemList/'+id).success(function(response){
-                refresh();
-            });
+        $scope.deleteItem = function(id, name){
+            if (confirm('Are you sure you want to delete '+name+'?')) {
+                $http.delete('/itemList/'+id).success(function(response){
+                    refresh();
+                });
+            }
         };
         $scope.selectItem = function(id){
             $http.get('/itemList/'+id).success(function(response){
@@ -31,16 +35,34 @@ angular.module('mainAppModule', ['loginModule'])
             })
         };
         $scope.clearItem = function(){
-            $scope.item = {
-                name:"", price:"0", qty:"0"
-            };
+            refreshInput();
         };
     })
     .controller("InitController2", function($scope, $http) {
 
-    });
+    })
+    .directive('numbersOnly', function () {
+    return {
+        require: 'ngModel',
+        link: function (scope, element, attr, ngModelCtrl) {
+            function fromUser(text) {
+                if (text) {
+                    var transformedInput = text.replace(/[^0-9]/g, '');
+
+                    if (transformedInput !== text) {
+                        ngModelCtrl.$setViewValue(transformedInput);
+                        ngModelCtrl.$render();
+                    }
+                    return transformedInput;
+                }
+                return undefined;
+            }
+            ngModelCtrl.$parsers.push(fromUser);
+        }
+    };
+});
 
 angular.module('loginModule', [])
     .controller("loginController", function($scope, $http) {
-        $scope.login='123456789';
+        $scope.login='loginModule';
     });
